requirements.txt
matplotlib
numpy
pandas
seaborn
scikit-learn
opencv-python
tensorflow-cpu

